In-game code editing
Now the text editor loads the example program at startup. You can edit, compile, and run it in real time. It controls one of the spheres with the move() function. This is a proof of concept that you can write code that interacts with the game world.
Also fixed a bug in the compiler.